+++ src/Debug/DebugReloader.py
@@ -21,7 +21,7 @@ else:
 class DebugReloader:
     def __init__(self, paths=None):
         if not paths:
-            paths = ["src", "plugins", config.data_dir + "/__plugins__"]
+            paths = ["src", "plugins"]
         self.log = logging.getLogger("DebugReloader")
         self.last_chaged = 0
         self.callbacks = []